Agents code duplicates removed
Inheriting from base class removes redundancy in the code across all the child class implementations. No warning from Pycharm now in agents.
