+++ agents/paired_state_agent.py
@@ -1,104 +1,32 @@
 import random
 import numpy as np
 
-from tensorflow.keras.models import Sequential
-from tensorflow.keras.layers import Dense
-from tensorflow.keras.optimizers import Adam
-from tensorflow.keras.models import model_from_json
-
-from annealers.paired_state_annealer import Annealer
-from utils.PER_memory_tree import Memory
+from agents.meta_agent import MetaDQNAgent
 
 
-class DQNAgent:
+class DQNAgent(MetaDQNAgent):
 
     def __init__(self, environment, memory_size=500):
-        self.environment = environment
-        self.furthest_distance = int(np.amax(self.environment.distance_matrix))
-        self.max_node_degree = int(np.max(np.sum(self.environment.adjacency_matrix, axis=1)))
-        self.memory_size = memory_size
+        super(DQNAgent, self).__init__(environment, memory_size)
 
-        # Set the Hyperparameters
+        # Set the Hyper-parameters
         self.gamma = 0.6
         self.epsilon = 1.0
         self.epsilon_min = 0.001
         self.epsilon_decay = 0.9
-        self.learning_rate = 0.001
-
         self.fix_learning_bug = True
 
-        self.NN_state_size = self.furthest_distance+1+self.max_node_degree+1
+        self.NN_state_size = self.furthest_distance + 1 + self.max_node_degree + 1
         self.current_model = self.build_model(self.NN_state_size)
         self.target_model = self.build_model(self.NN_state_size)
-
         self.update_target_model()
-
-        self.memory_tree = Memory(memory_size)
-        self.annealer = Annealer(self, environment)
 
     def build_model(self, furthest_distance):
         """
         Build the neural network model for this agent
         """
         input_size = furthest_distance * 2
-
-        model = Sequential([
-            Dense(32, input_dim=input_size, activation='relu'),
-            Dense(32, activation='relu'),
-            Dense(32, activation='relu'),
-            Dense(1, activation='linear'),
-        ])
-        model.compile(loss='mse',
-                      optimizer=Adam(lr=self.learning_rate))
-        return model
-
-    def update_target_model(self):
-        """
-        Copy weights from the current model to the target model
-        """
-        self.target_model.set_weights(self.current_model.get_weights())
-
-    def save_model(self, model_name=None):
-        # Serialize model to JSON
-        model_json = self.current_model.to_json()
-
-        if model_name is not None:
-            filepath = "./models/" + model_name
-        else:
-            filepath = "./models/agent_model"
-
-        with open(filepath + ".json", "w") as json_file:
-            json_file.write(model_json)
-
-        # Serialize weights to HDF5
-        self.current_model.save_weights(filepath + ".h5")
-        print("Saved model to disk")
-
-    def load_model(self, model_name=None):
-        self.epsilon = self.epsilon_min
-
-        if model_name is not None:
-            filepath = "./models/" + model_name
-        else:
-            filepath = "./models/agent_model"
-
-        # Load json and create model
-        json_file = open(filepath + '.json', 'r')
-        loaded_model_json = json_file.read()
-        json_file.close()
-        self.current_model = model_from_json(loaded_model_json)
-
-        # Load weights into new model
-        self.current_model.load_weights(filepath + ".h5")
-        self.current_model.compile(loss='mse', optimizer=Adam(lr=self.learning_rate))
-        self.update_target_model()
-        print("Loaded model from disk")
-
-    def remember(self, state, reward, next_state, done):
-        """
-        Store experience in the memory tree
-        """
-        self.memory_tree.store((state, reward, next_state, done))
+        return super().build_model(input_size)
 
     def generate_random_action(self, protected_nodes):
         """
@@ -116,12 +44,8 @@ class DQNAgent:
             edges = list(filter(lambda e: e[0] not in protected_nodes and e[1] not in protected_nodes, edges))
 
         while len(edges) > 0:
-            edge = random.sample(edges, 1)[0]
-            action[edge_index_map[edge]] = 1
-
-            # This also removes the sampled edge
+            edge, action[edge_index_map[edge]] = random.sample(edges, 1)[0], 1
             edges = [e for e in edges if e[0] not in edge and e[1] not in edge]
-
         return action
 
     def obtain_distance_vector(self, current_state):
@@ -267,9 +191,9 @@ class DQNAgent:
         qubit_targets = current_state[1]
 
         nodes_to_target_qubits = \
-            [qubit_targets[qubit_locations[n]] for n in range(0,len(qubit_locations))]
+            [qubit_targets[qubit_locations[n]] for n in range(0, len(qubit_locations))]
 
-        return np.reshape(np.array(nodes_to_target_qubits), (1,len(qubit_locations)))
+        return np.reshape(np.array(nodes_to_target_qubits), (1, len(qubit_locations)))
 
     def used_up_memory_capacity(self):
         return self.memory_tree.tree.used_up_capacity
